Remove getThemeSettingsSchemaForURI injection
Replace with abstract-file-system-based implementation

TODO make that shit faster. Seems highlighy unoptimized.
